\name{gnmf-package}
\alias{gnmf-package}
\docType{package}
\title{
Generalized non-negative matrix factorization
}
\description{
Generalized non-negative matrix factorization
}
\details{
\tabular{ll}{
Package: \tab gnmf\cr
Type: \tab Package\cr
Version: \tab 0.4\cr
Date: \tab 2012-6-3\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}
\author{
Jose M. Maisog, Guoli Wang, Karthik Devarajan
Maintainer: Jose M. Maisog <bravas02@gmail.com>
}
\references{
Devarajan K. Nonnegative matrix factorization: an analytical and interpretive 
tool in computational biology. PLoS Comput Biol. 2008 Jul 25;4(7):e1000029.

\code{\url{http://devarajan.fccc.edu}}
}
\keyword{ package }
\seealso{
\link{gnmf}
}
\examples{
data(V)

}
