\name{gnmf-package}
\alias{gnmf-package}
\docType{package}
\title{
Generalized Non-negative Matrix Factorization Based on Renyi Divergence
}
\description{
Provides a function \code{gnmf}, which performs generalized non-negative matrix factorization based on Renyi divergence
as described in Devarajan et al., 2011.
}
\details{
\tabular{ll}{
Package: \tab gnmf\cr
Type: \tab Package\cr
Version: \tab 0.4\cr
Date: \tab 2012-6-3\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
A function \code{gnmf} is provided to perform non-negative matrix factorization.
}
\author{
Jose M. Maisog, Guoli Wang, Karthik Devarajan
Maintainer: Jose M. Maisog <bravas02@gmail.com>
}
\references{
Devarajan K. Nonnegative matrix factorization: an analytical and interpretive 
tool in computational biology. PLoS Comput Biol. 2008 Jul 25;4(7):e1000029.

Devarajan, K., Wang, G.,  Ebrahimi, N. (2011). A unified approach to nonnegative matrix factorization and probabilistic latent semantic indexing, (July 2011). Cobra Preprint Series. Working Paper 80. \url{http://biostats.bepress.com/cobra/art80}.

URL: \code{http://devarajan.fccc.edu}
}
\keyword{ package }
\examples{
data(V)

}
